feat: Smaller modules across different crates
