+++ renderer/src/window.rs
@@ -1,7 +1,6 @@
 use dioxus_native_core::NodeId;
 use freya_common::EventMessage;
-use freya_core::process_render;
-use freya_core::{process_layout, ViewportsCollection};
+use freya_core::prelude::*;
 use freya_dom::prelude::FreyaDOM;
 use freya_layout::Layers;
 use std::ffi::CString;
